Change version from Long to BigInteger

+++ src/main/java/uk/gov/hmcts/opal/dto/DefendantAccountResponse.java
@@ -3,6 +3,7 @@ package uk.gov.hmcts.opal.dto;
 import com.fasterxml.jackson.annotation.JsonIgnore;
 import com.fasterxml.jackson.annotation.JsonInclude;
 import com.fasterxml.jackson.annotation.JsonProperty;
+import java.math.BigInteger;
 import lombok.AllArgsConstructor;
 import lombok.Builder;
 import lombok.Data;
@@ -38,6 +39,6 @@ public class DefendantAccountResponse implements ToJsonString, Versioned {
     private EnforcementOverride enforcementOverride;
 
     @JsonIgnore
-    private Long version;
+    private BigInteger version;
 
 }
